refactor: Removes tasks-organize-multiple in favour of update
